Cleanup for tests files
