--- apps/evm/tests/vanchor/fungible-token-wrapper.test.ts
@@ -1,21 +1,0 @@
-import { beforeAll, beforeEach, clearStore, describe } from "matchstick-as/assembly/index";
-
-// Tests structure (matchstick-as >=0.5.0)
-// https://thegraph.com/docs/en/developer/matchstick/#tests-structure-0-5-0
-
-describe('Fungible token wrapper transaction', () => {
-
-  beforeEach(() =>{
-    clearStore();
-
-  })
-  beforeAll(() => {
-    clearStore();
-  });
-
-
-});
-/// test
-0x0000000000000000000000000000000000000000000000000000000000000020000000000000000000000000000000000000000000000000000000000000000500000000000000000000000000000000000000000000000000000000000000a000000000000000000000000000000000000000000000000000000000000001c0000000000000000000000000000000000000000000000000000000000000020000000000000000000000000000000000000000000000000000000000000002e000000000000000000000000000000000000000000000000000000000000007a000000000000000000000000000000000000000000000000000000000000001000475bdaa274a1d55391cfcf587afefadc4c497c70b6d0a62e1b51154ed5969bb26642b4c5ee9f6e1bdfd4d8fa06a600699c1a2599197b6fa82b95b3a0ce58779000a1f1af2b452e0f0342719a0777f5b0e3cfdc3e261b62fa971b9a247bb6ece0500aea30437d112bd5cd8e5bf8cfb65418ebb41b642d371c6ea49b555450f6d069721c72857cf2f30b931f2eed6e4665f4053fbbc1a91932e4d58343ece88e32617430f4ffac8617e1a380afa6ce4f0c5244d0f916813538808ce9778d3cabd1b44b57c47b4939020ea973b8e71e375a65bee652727a5bbf14d556c7dce9689038cf25dd97a23785a5131c27d2daf3e340c40afad489da83a588c182bbc26160000000000000000000000000000000000000000000000000000000000000020000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000060000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000001e0000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000600000000000000000000000000000000000000000000000000000000000000c000000000000000000000000000000000000000000000000000000000000001e000000000000000000000000000000000000000000000000000000000000002000000000000000000000000000000000000000000000000000000000000000320000000000000000000000000000000000000000000000002b5e3af16b1880000000000000000000000000000000000000000000000000000000000000000044000000000000000000000000000000000000000000000000000000000000001001d15b12367fb5fcc9739bd8a70aab1a59f4d3c96281a40b2ca0e4925b5e5692c0cebe3de811c0e6e3466cd1a463c759b4019dc19d5b24348be425730f63db9dc23ab323453748129f2765f79615022f5bebd6f4096a796300aab049a60b0f18723ab323453748129f2765f79615022f5bebd6f4096a796300aab049a60b0f18723ab323453748129f2765f79615022f5bebd6f4096a796300aab049a60b0f18723ab323453748129f2765f79615022f5bebd6f4096a796300aab049a60b0f18723ab323453748129f2765f79615022f5bebd6f4096a796300aab049a60b0f18723ab323453748129f2765f79615022f5bebd6f4096a796300aab049a60b0f18700000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000002000000000000000000000000000000000000000000000000000000000000004000000000000000000000000000000000000000000000000000000000000000a00000000000000000000000000000000000000000000000000000000000000027212268307691617886401607482596124063880320050222970563349910327581577426076981000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000027135871326026479886979432225270531260060135152626859453912057339975698563302611000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000002000000000000000000000000000000000000000000000000000000000000004000000000000000000000000000000000000000000000000000000000000000a000000000000000000000000000000000000000000000000000000000000000271228035841339105195481598203448104642110171751544113653526783796280136059068510000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000271775312328822592487133529285834109054824600149083720647724390289847479311339410000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000265851782729587194271435774257001525337865823128641196509784354335913672864280000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000020000000000000000000000000000000000000000000000000000000000000040000000000000000000000000000000000000000000000000000000000000012000000000000000000000000000000000000000000000000000000000000000a856dc89653fd49ee105c84cc2d6b6bae5362cac2aff8fbbd2765ae912d6450c9d2298ccd25053133c6e08e9e3c773b8496baf8ea3de3d12055e3a3183bc50d83337956849df1263dcad79bba7512e8166d04df05cf4fcf2dc1b35008452f3e98b626c1bea2b0953324c96380eb9a5f6e9b59822eb6a8330232ec524522da3a4de8ef90fc8cb257cbe5d01e17a147d229466c821ef475f03ce55ddac216160a42e1f9f89ff8c8485f600000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000a813afe445ecbcd4e16b1e4bef4a32f0748d38b8e459c91645be225e457bce1d8e72f46d66452f943c765d59976d2dff8d5f900239d2c568781b4e9f34bb0cbfe4445019a5707214d2cb1261fc0ed8b118aab38aace9180acb1ebb98da563f1c732568f276f031f7413218fb2a470bb98de3676e8d70677b1a67b5e1776a051400452dd28eb143149ba1b09faa96f12480b45b654bf66b2b07d1daf763fa54c4d23632a5abea967a60000000000000000000000000000000000000000000000000
-/// one
-'0x000000000000000000000000000000000000000000000000000000000000002000000000000000000000000000000000000000000000000000000000000001400000000000000000000000000000000000000000000000000000000000000260000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000002540be400000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000fdaacacb85484c4d492414f4911524f57d2549f400000000000000000000000000000000000000000000000000000000000002a000000000000000000000000000000000000000000000000000000000000005200000000000000000000000000000000000000000000000000000000000000100154ccb9bb95c88df32d14db483224da8a8abbda31fe88e199965471c48b7f05d19ae0dc768c0b19241ed974fb64070bc53f5c86bdf583da1f34ff483c2a04674287a4313e55e5f256b3038943c30a2fb92c48659f1e501feb14b72986e7cd4ed266ea2156db02ef96d8eacb1e3c7c7a8fa3dad43d20cb06003aa686cce068eff156aa9dbeb1a3a8c715c6f4651b562e1509ff3fe1266ab9fbcd4d0e7d2eaa00f1e91de6d4f3fa83eef75f5e9a851d09a7e5b451c1eebd1df4ff8c33958e5e10a04979f104fe45408ec76fceb01578a1c190504e6597e88ea6cb0bf2b13b53ea70eccb72cc67027d55590317be34f266ca878bba38a59d12b45ccebcccfcb2c2f0000000000000000000000000000000000000000000000000000000000000020000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000e0000000000000000000000000000000000000000000000000000000000000020000000000000000000000000000000000000000000000000000000000000002202924b18add75081fab7d484df212f23f4d136b70bfa493d1d5b5b5a5796039fa1533037e471a2fed2ec7a37e58de34a5a2cbee422b33112988f5f9212a26684600000000000000000000000000000000000000000000000000000002540be4001c8bb4acc4429433d44376528b26d2d51591081fe395eb74cdd4c042a024ea90000000000000000000000000000000000000000000000000000000000000010023ab323453748129f2765f79615022f5bebd6f4096a796300aab049a60b0f18723ab323453748129f2765f79615022f5bebd6f4096a796300aab049a60b0f18723ab323453748129f2765f79615022f5bebd6f4096a796300aab049a60b0f18723ab323453748129f2765f79615022f5bebd6f4096a796300aab049a60b0f18723ab323453748129f2765f79615022f5bebd6f4096a796300aab049a60b0f18723ab323453748129f2765f79615022f5bebd6f4096a796300aab049a60b0f18723ab323453748129f2765f79615022f5bebd6f4096a796300aab049a60b0f18723ab323453748129f2765f79615022f5bebd6f4096a796300aab049a60b0f1870000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000213104d767bf2a2354490f86a0334ec890359f7c7e1ebd8ef652dd88fe5ef99ef214135a5c5a83e853814b3c56d40712ade4cf5a7d8f1b38dd47575373ee4f1120000000000000000000000000000000000000000000000000000000000000040000000000000000000000000000000000000000000000000000000000000012000000000000000000000000000000000000000000000000000000000000000a8623b0fa1a2ff0387b07cd6230d8ad1e3534fdddfdd033d80a266adafb2cc4c8abfee0bcf5d8040bd91239fb5f1c642add232dd501507ac1b3174e87e2769901664c250fe1994cd6fea35163ca265c027fc54f444750433581617b419b7811304e064c248a3e66c3a442bc4bcb96694ca8005e5d1e06d246fdec4eb14799f3a375bdd38891fc9da057a43bfbaa465a100259a7247e11c9e3af3d91e8bb296bee96a0ee4deb02cb4bc00000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000a892abcaacbbd867fabed9b63cf25b48233ffcdcd60f33ccc55fa277256c956bccb39f603be8b2731f5d1572bce5f7257bec0d25049de1f020107fd744324db887543a7e7470f650fb48b381c0c97d9a221edc0e43767ffe790b103abf0a6cdd948f08b25d9708032935e17da4e34374930db7758df9c0f4627242821217c034dc6b0d8d0acbf2557877fefb2563e6954af416664ebd891a53f2cb15bbef8070bd745baad2663d3589000000000000000000000000000000000000000000000000'
